# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: HomeStock
  description: '

    A self-hosted home inventory system.


    Every response uses the same envelope: `{"data": ..., "error": null}`.

    On failure `data` is null and `error` holds a code and a message.


    The server runs every AI model on the CPU. There is no GPU. The AI routes

    return a job instead of a result when inference takes longer than

    `HS_AI_INLINE_TIMEOUT` seconds.

    '
  version: 0.1.0
paths:
...
  /api/items/{item_id}/activity:
    get:
      tags:
      - Activity
      summary: Return the log of one item. The newest line comes first.
      operationId: item_activity_api_items__item_id__activity_get
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_ActivityRow__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/activity:
    get:
      tags:
      - Activity
      summary: Return the newest lines of the whole log.
      operationId: recent_activity_api_activity_get
      security:
      - HTTPBearer: []
      parameters:
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Only this kind of line.
          title: Action
        description: Only this kind of line.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_ActivityRow__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/cables:
    get:
      tags:
      - Cables
      summary: List the cables.
      description: '``connector`` matches either end, because you look for a cable with USB-C on it and
        not for a direction.'
      operationId: list_cables_api_cables_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Match the name, the brand, or the ends.
          title: Q
        description: Match the name, the brand, or the ends.
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      - name: connector
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Match either end.
          title: Connector
        description: Match either end.
      - name: location_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
      - name: item_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_CableRead__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Cables
      summary: Add one cable.
      operationId: create_cable_api_cables_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CableCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_CableRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/cables/kinds:
    get:
      tags:
      - Cables
      summary: List the cable families that are in use.
      operationId: list_kinds_api_cables_kinds_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_list_str__'
      security:
      - HTTPBearer: []
  /api/cables/{cable_id}:
    get:
      tags:
      - Cables
      summary: Return one cable.
      operationId: get_cable_api_cables__cable_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: cable_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Cable Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_CableRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Cables
      summary: Change a cable.
      operationId: update_cable_api_cables__cable_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: cable_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Cable Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CableUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_CableRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Cables
      summary: Delete a cable.
      operationId: delete_cable_api_cables__cable_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: cable_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Cable Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Message_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActivityRow:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: string
          format: uuid
          title: Entity Id
        action:
          type: string
          title: Action
        summary:
          type: string
          title: Summary
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        actor:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - entity_type
      - entity_id
      - action
      - summary
      - detail
      - actor
      - version
      - created_at
      - updated_at
      title: ActivityRow
      description: One line of the log.
...
    CableCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        kind:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Kind
          description: 'The family: "USB", "HDMI", "Ethernet", or "Power".'
        connector_a:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Connector A
        connector_b:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Connector B
        length_cm:
          anyOf:
          - type: integer
            maximum: 100000.0
            minimum: 0.0
          - type: 'null'
          title: Length Cm
          description: The length in centimetres.
        colour:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Colour
        brand:
          anyOf:
          - type: string
            maxLength: 150
          - type: 'null'
          title: Brand
        specification:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Specification
          description: 'What it can do: "USB 3.2, 10 Gbps, 100 W" or "Cat6a".'
        quantity:
          type: integer
          minimum: 0.0
          title: Quantity
          default: 1
        price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
          - type: 'null'
          title: Price
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
          description: The device that this lead came with.
      type: object
      required:
      - name
      title: CableCreate
      description: Add one cable.
    CableRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        connector_a:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector A
        connector_b:
          anyOf:
          - type: string
          - type: 'null'
          title: Connector B
        length_cm:
          anyOf:
          - type: integer
          - type: 'null'
          title: Length Cm
        colour:
          anyOf:
          - type: string
          - type: 'null'
          title: Colour
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        specification:
          anyOf:
          - type: string
          - type: 'null'
          title: Specification
        quantity:
          type: integer
          title: Quantity
        price:
          anyOf:
          - type: string
          - type: 'null'
          title: Price
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        ends:
          anyOf:
          - type: string
          - type: 'null'
          title: Ends
          description: Such as "USB-C to HDMI".
        length_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Length Label
          description: Such as "2 m".
        total_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Total Value
          description: The price of one, times the count.
        location_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Name
        item_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Item Name
        thumbnail_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Path
          description: The picture that stands for this cable.
        photo_count:
          type: integer
          title: Photo Count
          default: 0
      type: object
      required:
      - id
      - user_id
      - name
      - kind
      - connector_a
      - connector_b
      - length_cm
      - colour
      - brand
      - specification
      - quantity
      - price
      - notes
      - location_id
      - item_id
      - version
      - created_at
      - updated_at
      title: CableRead
      description: One cable, with the words that the interface shows.
...
    CableUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        kind:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Kind
        connector_a:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Connector A
        connector_b:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Connector B
        length_cm:
          anyOf:
          - type: integer
            maximum: 100000.0
            minimum: 0.0
          - type: 'null'
          title: Length Cm
        colour:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Colour
        brand:
          anyOf:
          - type: string
            maxLength: 150
          - type: 'null'
          title: Brand
        specification:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Specification
        quantity:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Quantity
        price:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
          - type: 'null'
          title: Price
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        location_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Location Id
        item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Item Id
      type: object
      title: CableUpdate
      description: Change a cable. Every field is optional.
...
    Envelope_CableRead_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/CableRead'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[CableRead]
...
    Envelope_Message_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Message'
          - type: 'null'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[Message]
...
    Envelope_list_ActivityRow__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/ActivityRow'
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[list[ActivityRow]]
    Envelope_list_CableRead__:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/CableRead'
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[list[CableRead]]
...
    Envelope_list_str__:
      properties:
        data:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Data
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorDetail'
          - type: 'null'
      type: object
      title: Envelope[list[str]]
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          description: A stable machine readable code.
        message:
          type: string
          title: Message
          description: A message for a person to read.
        details:
          anyOf:
          - type: object
          - type: 'null'
          title: Details
          description: Extra context, such as field errors.
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: The error half of the response envelope.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: Message
      description: A plain acknowledgement, for routes that return no object.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      description: A JWT access token.
      scheme: bearer